The HTML stylesheet likes to print html has the ends of tags on a different line, like this: <HTML ><HEAD ><TITLE >FreeBSD</TITLE >
Glimpse, which is indexing our website, finds this very confusing and
it cannot pick out the title from this mess.  This script takes a list
of HTML files on the command line and attempts to make the <TITLE> tag
look more normal so that glimpse can understand it.

WARNING: This is a hack.  It's made to work on docbook generated html, but
may do strange things on anything else.